+++ src/error/index.h
@@ -2,9 +2,10 @@
 #define TIGER_ERROR
 
 bool hasErrors();
-void reportError(const char * message);
+void reportError(const char* message);
 void reportBadToken();
 void refresh();
+void yyerror(const char*);
 
 #endif