+++ src/syntax/dump.h
@@ -1,8 +1,0 @@
-#ifndef TIGER_SYNTAX_DUMP
-#define TIGER_SYNTAX_DUMP
-
-namespace syntax {
-    void printExpr(struct Expr *expr);
-}
-
-#endif