+++ test/ast.spec.cc
@@ -1,0 +1,39 @@
+#include <assert.h>
+#include <stdio.h>
+#include "../src/utils/index.h"
+#include "../src/lexical/index.h"
+#include "../src/ast/index.h"
+#include "global.h"
+
+extern int yyparse(void);
+
+namespace ast {
+    int parse(const char *fname) {
+        lexical::reset(fname);
+        if (yyparse() == 0) {
+            return 0;
+        } else {
+            return 1;
+        }
+    }
+
+    void debug() {
+        printf("\n~~~~~~ DEBUGGING AST STARTED ~~~~~~\n\n");
+        assert(ast::parse(TIGER_DEFAULT_INPUT) == 0);
+        // ast::AST_ROOT->print();
+        printf("\n\n~~~~~~ DEBUGGING AST ENDED ~~~~~~\n");
+    }
+}
+
+static int __TIGER_UNIT_TEST = describe("ast tree", [] {
+
+    it("should parse ast correct", [] {
+        for (auto i: TIGS) {
+            if (i == "../mocks/merge.tig") continue;
+            if (i == "../mocks/queens.tig") continue;
+            assert(ast::parse(i.c_str()) == 0);
+        }
+
+        // ast::debug(); /* if want observe */
+    });
+});